requirements_dashboard.txt
streamlit
pandas
psycopg2-binary
plotly